Edit category form moved into new format
